requirements.txt
adjustText==1.3.0
beautifulsoup4==4.13.3
googlesearch_python==1.3.0
json5==0.10.0
langchain==0.3.19
langchain_google_genai==2.0.10
matplotlib==3.10.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.3
openai==1.64.0
pandas==2.2.3
protobuf==5.29.3
pyvis==0.3.2
PyYAML==6.0.2
Requests==2.32.3
requests_html==0.10.0
scipy==1.15.2
selenium==4.29.0
tqdm==4.67.1
webdriver_manager==4.0.2
wordcloud==1.9.4